+++ src/Migrations/AegisIdentity.Migrations/EfMigrations/20260614220101_AddUserIsBootstrapColumn.Designer.cs
@@ -4,6 +4,7 @@ using AegisIdentity.DataAccess.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 #nullable disable
@@ -11,9 +12,11 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace AegisIdentity.Migrations.EfMigrations
 {
     [DbContext(typeof(AegisIdentityDbContext))]
-    partial class AegisIdentityDbContextModelSnapshot : ModelSnapshot
+    [Migration("20260614220101_AddUserIsBootstrapColumn")]
+    partial class AddUserIsBootstrapColumn
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder