+++ src/Presentation/AegisIdentity.Backoffice/Controllers/HomeController.cs
@@ -1,4 +1,5 @@
 using AegisIdentity.Backoffice.Services;
+using AegisIdentity.Backoffice.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -35,30 +36,30 @@ public sealed class HomeController : BackofficeBaseController
         {
         }
 
-        var userCount = await SafeResult(userCountTask);
-        if (userCount.HasValue) ViewBag.UserCount = userCount.Value;
-
-        var profiles = await SafeResult(profilesTask);
-        if (profiles is not null) ViewBag.ProfileCount = profiles.Count;
-
+        var userCount        = await SafeResult(userCountTask);
+        var profiles         = await SafeResult(profilesTask);
         var permissionGroups = await SafeResult(permissionsTask);
+        var activity         = await SafeResult(activityTask);
+        var cacheStats       = await SafeResult(cacheStatsTask);
+        var jobStats         = await SafeResult(jobStatsTask);
+
+        int? permissionCount = null;
+        int? orphanCount = null;
         if (permissionGroups is not null)
         {
             var all = permissionGroups.SelectMany(g => g.Permissions).ToList();
-            ViewBag.PermissionCount = all.Count;
-            ViewBag.OrphanCount     = all.Count(p => p.IsOrphan);
+            permissionCount = all.Count;
+            orphanCount     = all.Count(p => p.IsOrphan);
         }
 
-        var activity = await SafeResult(activityTask);
-        if (activity is not null) ViewBag.Activity = activity;
-
-        var cacheStats = await SafeResult(cacheStatsTask);
-        if (cacheStats?.HitRate is not null) ViewBag.CacheHitRate = cacheStats.HitRate;
-
-        var jobStats = await SafeResult(jobStatsTask);
-        if (jobStats is not null) ViewBag.JobStats = jobStats;
-
-        return View();
+        return View(new HomeDashboardViewModel(
+            UserCount:       userCount,
+            ProfileCount:    profiles?.Count,
+            PermissionCount: permissionCount,
+            OrphanCount:     orphanCount,
+            CacheHitRate:    cacheStats?.HitRate,
+            Activity:        activity,
+            JobStats:        jobStats));
     }
 
     private static async Task<T?> SafeResult<T>(Task<T?> task) where T : class